casting to double then getting int value to reduce errors when input comes as a string with a decimal value of 0
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@14013 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 232e69c8a4fcf361f7a12532bb5d8545e0d9efb3

+++ src/prerna/engine/impl/QuestionAdministrator.java
@@ -113,7 +113,8 @@ public class QuestionAdministrator {
 		if(order == null || order.isEmpty()) {
 			order = calculateDefaultOrdering(perspective) + "";
 		} else {
-			cleanPerspectiveOrdering(perspective, Integer.parseInt(order));
+			// cast as double and then get integer value due to examples when string comes back as "5.0"
+			cleanPerspectiveOrdering(perspective, (int) Double.parseDouble(order));
 		}
 		
 		// insert into table the new record
